pyproject.toml
[tool.poetry]
name = "Crawler"
version = "0.1.0"
description = "FastAPI project generated by script"
authors = ["Anh Ng. <nvanh.vnu@gmail.com>"]
license = "MIT" # Example license
readme = "README.md"
packages = [ # Define packages for non-src layout
    { include = "." }
]
# homepage = ""
# repository = ""
# documentation = ""
# keywords = ["fastapi", "ddd"]

[tool.poetry.dependencies]
python = "^3.10"
# Core dependencies will be added here by 'poetry add' below
fastapi = "^0.124.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
alembic = "^1.17.2"
httpx = "^0.28.1"
redis = {extras = ["hiredis"], version = "^7.1.0"}
grpcio = "^1.76.0"
grpcio-tools = "^1.76.0"
email-validator = "^2.3.0"
python-json-logger = "^4.0.0"
psycopg2-binary = "^2.9.11"
aiosqlite = "^0.21.0"
asyncpg = "^0.31.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
# Development dependencies will be added here by 'poetry add --group dev' below
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
httpx = "^0.28.1"
mypy = "^1.19.0"
ruff = "^0.14.8"
pre-commit = "^4.5.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.5"
types-redis = "^4.6.0.20241004"
types-passlib = "^1.7.7.20250602"
types-python-jose = "^3.5.0.20250531"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# --- Optional: Ruff Configuration ---
[tool.ruff]
line-length = 88
# select = ["E", "W", "F", "I", "UP", "PL", "T20"] # Example rule selection
# ignore = ["E501"]

# --- Optional: MyPy Configuration ---
[tool.mypy]
python_version = "3.10"
# warn_return_any = true
# warn_unused_configs = true
# Add other mypy options

